fix: make customer 360 validator executable
