+++ testing/ci/__init__.py
@@ -1,0 +1,1 @@
+"""CI validation entry points."""
